Unit test name refactoring
